Server commands moved into core-command plugin and beginning of migrating time/weather plugin
