+++ src/core/plugins/core-commands/server/commands/player/vehicle.ts
@@ -1,12 +1,11 @@
 import * as alt from 'alt-server';
-
-import { PERMISSIONS } from '../../shared/flags/permissionFlags';
-import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
-import { LocaleController } from '../../shared/locale/locale';
-import { Athena } from '../api/athena';
-import { command } from '../decorators/commands';
-import VehicleFuncs from '../extensions/vehicleFuncs';
-import { VehicleSystem } from '../systems/vehicle';
+import { Athena } from '../../../../../server/api/athena';
+import { command } from '../../../../../server/decorators/commands';
+import VehicleFuncs from '../../../../../server/extensions/vehicleFuncs';
+import { VehicleSystem } from '../../../../../server/systems/vehicle';
+import { PERMISSIONS } from '../../../../../shared/flags/permissionFlags';
+import { LOCALE_KEYS } from '../../../../../shared/locale/languages/keys';
+import { LocaleController } from '../../../../../shared/locale/locale';
 
 class VehicleCommands {
     @command('engine', LocaleController.get(LOCALE_KEYS.COMMAND_TOGGLE_ENGINE, '/engine'), PERMISSIONS.NONE)
