--- src/core/server/commands/job.ts
@@ -1,29 +1,0 @@
-import * as alt from 'alt-server';
-import { PERMISSIONS } from '../../shared/flags/permissionFlags';
-import { LOCALE_KEYS } from '../../shared/locale/languages/keys';
-import { LocaleController } from '../../shared/locale/locale';
-import { Athena } from '../api/athena';
-import ChatController from '../systems/chat';
-import { getPlayerJob } from '../systems/job';
-
-ChatController.addCommand(
-    'quitjob',
-    LocaleController.get(LOCALE_KEYS.COMMAND_QUIT_JOB, '/quitjob'),
-    PERMISSIONS.NONE,
-    handleCommand,
-);
-
-function handleCommand(player: alt.Player): void {
-    if (!player || !player.valid) {
-        return;
-    }
-
-    const job = getPlayerJob(player);
-
-    if (!job) {
-        Athena.player.emit.notification(player, LocaleController.get(LOCALE_KEYS.JOB_NOT_WORKING));
-        return;
-    }
-
-    job.quit(LocaleController.get(LOCALE_KEYS.JOB_QUIT));
-}
